jpa sett upp, en á eftir að tengja saman tables. signup komið, ekki login samt

+++ src/main/java/hi/hugboverkefni1/services/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface UserService {
 
     List<User> getUsers();
-    User findUserById(int id);
+    User findUserById(long id);
     void saveUser(User user);
     void deleteUser(User user);
